+++ core/decorators/test-cases-decorator.ts
@@ -1,6 +1,7 @@
 import "reflect-metadata";
-import { TEST_CASES, TESTS } from "./_metadata-keys";
+import { TEST_CASES } from "./_metadata-keys";
 import { Unused } from "../unused";
+import { markPropertyAsTest } from "./mark-property-as-test";
 
 function GetTestCases(
   caseArguments:
@@ -40,24 +41,7 @@ export function TestCases(
   ) => {
     Unused(descriptor);
 
-    // check if this has been registered as a test already
-    let tests: Array<any> = Reflect.getMetadata(TESTS, target);
-
-    if (!tests) {
-      // if there are no tests registered yet then register it
-      tests = [
-        {
-          key: propertyKey
-        }
-      ];
-      Reflect.defineMetadata(TESTS, tests, target);
-    } else if (tests.filter(test => test.key === propertyKey).length === 0) {
-      // otherwise add it to the register
-      tests.push({
-        key: propertyKey
-      });
-      Reflect.defineMetadata(TESTS, tests, target);
-    }
+    markPropertyAsTest(propertyKey, target);
 
     const testCases = GetTestCases(caseArguments).reduce((acc, val) => {
       // add the test case to the list
